[fix] Bump version (1.1.2), small changes
This is a fix, the script didn't run as expected on version 1.1.0.

[bbackup.sh]
Fixed some typos and changed the execution order.
Added a '$_execpath' variable so the libraries get included no matter where the script get's executed.
Checking for a running instance is now the first thing before anything gets loaded and configured.
Fixed the '$TASK' and '$CONFIG_FILE' error -> "Unbound variable ...".
Checking for compression flag before trying to calculate it's size.

[bbackup.default.conf]
Changed the "TASK" from the config to "DEFAULT_TASK" for better assignment and cleaner code.

[log.sh]
Fixed a typo in the 'printf' statement.
Made the exposed variables global, so they can be modified from bbackup.sh.

[notificatino.sh]
Changed the '$JOB' variable to '$TASK' in the 'mail_str' variable.